Moves halt_play files to their own directory and adds the necessary files
