+++ src/software/ai/BUILD
@@ -18,7 +18,7 @@ cc_library(
         "//proto:play_info_msg_cc_proto",
         "//software/ai/hl/stp/play:all_plays",
         "//software/ai/hl/stp/play:assigned_tactics_play",
-        "//software/ai/hl/stp/play:halt_play",
+        "//software/ai/hl/stp/play/halt_play:halt_play",
         "//software/ai/hl/stp/play:play_factory",
         "//software/ai/hl/stp/tactic:tactic_factory",
         "//software/time:timestamp",
