+++ src/software/ai/hl/stp/play/BUILD
@@ -20,19 +20,7 @@ cc_library(
     alwayslink = True,
 )
 
-cc_library(
-    name = "halt_play",
-    srcs = ["halt_play.cpp"],
-    hdrs = ["halt_play.h"],
-    deps = [
-        ":play",
-        "//shared:constants",
-        "//software/ai/hl/stp/tactic/stop:stop_tactic",
-        "//software/logger",
-        "//software/util/generic_factory",
-    ],
-    alwayslink = True,
-)
+
 
 cc_library(
     name = "kickoff_enemy_play",
@@ -134,7 +122,7 @@ cc_library(
     name = "all_plays",
     deps = [
         ":example_play",
-        ":halt_play",
+        "//software/ai/hl/stp/play/halt_play:halt_play",
         ":kickoff_enemy_play",
         ":kickoff_friendly_play",
         ":shoot_or_chip_play",
@@ -234,37 +222,9 @@ cc_test(
     ],
 )
 
-cc_test(
-    name = "halt_play_cpp_test",
-    srcs = ["halt_play_test.cpp"],
-    deps = [
-        "//shared/test_util:tbots_gtest_main",
-        "//software/ai/hl/stp/play:halt_play",
-        "//software/simulated_tests:simulated_er_force_sim_play_test_fixture",
-        "//software/simulated_tests/terminating_validation_functions",
-        "//software/simulated_tests/validation:validation_function",
-        "//software/test_util",
-        "//software/time:duration",
-        "//software/world",
-    ],
-)
 
-py_test(
-    name = "halt_play_test",
-    srcs = [
-        "halt_play_test.py",
-    ],
-    # TODO (#2619) Remove tag to run in parallel
-    tags = [
-        "exclusive",
-    ],
-    deps = [
-        "//software:conftest",
-        "//software/simulated_tests:speed_threshold_helpers",
-        "//software/simulated_tests:validation",
-        requirement("pytest"),
-    ],
-)
+
+
 
 cc_test(
     name = "shoot_or_chip_play_cpp_test",
