+++ src/software/ai/hl/stp/stp_referee_command_play_selection_test.cpp
@@ -3,7 +3,7 @@
 #include <exception>
 
 #include "proto/parameters.pb.h"
-#include "software/ai/hl/stp/play/halt_play.h"
+#include "software/ai/hl/stp/play/halt_play/halt_play.h"
 #include "software/ai/hl/stp/stp.h"
 #include "software/test_util/test_util.h"
 #include "software/world/world.h"
