+++ src/software/ai/hl/stp/stp_test.cpp
@@ -5,7 +5,7 @@
 #include <algorithm>
 #include <exception>
 
-#include "software/ai/hl/stp/play/halt_play.h"
+#include "software/ai/hl/stp/play/halt_play/halt_play.h"
 #include "software/test_util/test_util.h"
 
 class STPTest : public ::testing::Test
