+++ src/software/ai/play_selection_fsm.cpp
@@ -4,7 +4,7 @@
 #include "software/ai/hl/stp/play/enemy_ball_placement/enemy_ball_placement_play.h"
 #include "software/ai/hl/stp/play/enemy_free_kick/enemy_free_kick_play.h"
 #include "software/ai/hl/stp/play/free_kick/free_kick_play.h"
-#include "software/ai/hl/stp/play/halt_play.h"
+#include "software/ai/hl/stp/play/halt_play/halt_play.h"
 #include "software/ai/hl/stp/play/kickoff_enemy_play.h"
 #include "software/ai/hl/stp/play/kickoff_friendly_play.h"
 #include "software/ai/hl/stp/play/offense/offense_play.h"
