+++ src/software/ai/play_selection_fsm_test.cpp
@@ -4,7 +4,7 @@
 
 #include <exception>
 
-#include "software/ai/hl/stp/play/halt_play.h"
+#include "software/ai/hl/stp/play/halt_play/halt_play.h"
 #include "software/test_util/test_util.h"
 #include "software/world/world.h"
