+++ src/software/simulated_tests/BUILD
@@ -44,7 +44,7 @@ cc_library(
         "//shared/test_util:tbots_gtest_main",
         "//shared/test_util",
         # TODO (#1889) Remove this dep after optional params are implemented
-        "//software/ai/hl/stp/play:halt_play",
+        "//software/ai/hl/stp/play/halt_play:halt_play",
         "//proto/message_translation:er_force_world",
     ],
 )
