+++ src/software/simulated_tests/simulated_er_force_sim_test_fixture.h
@@ -2,7 +2,7 @@
 
 #include "proto/play_info_msg.pb.h"
 #include "shared/test_util/tbots_gtest_main.h"
-#include "software/ai/hl/stp/play/halt_play.h"
+#include "software/ai/hl/stp/play/halt_play/halt_play.h"
 #include "software/sensor_fusion/sensor_fusion.h"
 #include "software/simulated_tests/validation/non_terminating_function_validator.h"
 #include "software/simulated_tests/validation/terminating_function_validator.h"
